Extraced code in separate component with local state
