+++ src/components/Chat.tsx
@@ -1,13 +1,13 @@
 ﻿import * as React from 'react';
-import {connect} from 'react-redux';
-import {Action, Dispatch} from 'redux';
-import EmojiPicker, {EmojiClickData, EmojiStyle, SkinTones} from 'emoji-picker-react';
+import { connect } from 'react-redux';
+import { Action, Dispatch } from 'redux';
+import { EmojiClickData } from 'emoji-picker-react';
 import State from '../state/State';
 import Constants from '../model/enums/Constants';
 import actionCreators from '../state/actionCreators';
 import ChatMessage from '../model/ChatMessage';
 import ChatLog from './common/ChatLog';
-
+import ChatInputEmojiPicker from './common/ChatInputEmojiPicker';
 import './Chat.css';
 
 interface ChatOwnProps {
@@ -21,12 +21,7 @@ interface ChatStateProps {
 	messages: ChatMessage[];
 }
 
-interface ChatInputEmojiPickerProps {
-	isEmojiPickerOpened: boolean;
-	onEmojiPickerToggle: (isOpened: boolean) => void;
-}
-
-interface ChatProps extends ChatStateProps, ChatOwnProps, ChatInputEmojiPickerProps {
+interface ChatProps extends ChatStateProps, ChatOwnProps {
 
 }
 
@@ -34,7 +29,6 @@ const mapStateToProps = (state: State) => ({
 	isConnected: state.common.isConnected,
 	currentMessage: state.online.currentMessage,
 	messages: state.online.messages,
-	isEmojiPickerOpened: state.online.isEmojiPickerOpened,
 });
 
 const mapDispatchToProps = (dispatch: Dispatch<Action>) => ({
@@ -43,9 +37,6 @@ const mapDispatchToProps = (dispatch: Dispatch<Action>) => ({
 	},
 	onSendMessage: () => {
 		dispatch((actionCreators.sendMessage() as object) as Action);
-	},
-	onEmojiPickerToggle: (isOpened: boolean) => {
-		dispatch(actionCreators.onEmojiPickerToggle(isOpened));
 	}
 });
 
@@ -63,11 +54,6 @@ export function Chat(props: ChatProps): JSX.Element {
 			e.preventDefault();
 		}
 	};
-
-	const onEmojiPickerButtonClick = (e: React.MouseEvent<HTMLButtonElement>) => {
-		props.onEmojiPickerToggle(!props.isEmojiPickerOpened)
-	};
-
 	const onEmojiClick = (emojiData: EmojiClickData, e: MouseEvent) => {
 		props.onMessageChanged(props.currentMessage + emojiData.emoji);
 	};
@@ -75,7 +61,7 @@ export function Chat(props: ChatProps): JSX.Element {
 	return (
 		<div className="chatBodyHost">
 			<ChatLog className="chat" messages={props.messages} />
-
+			<ChatInputEmojiPicker onEmojiClick={onEmojiClick} />
 			<input
 				type='text'
 				className={`message ${props.isConnected ? '' : 'disconnected'}`}
@@ -83,29 +69,6 @@ export function Chat(props: ChatProps): JSX.Element {
 				onChange={onMessageChanged}
 				onKeyPress={onMessageKeyPress}
 			/>
-
-			<button
-				type={'button'}
-				className={'standard chatEmojiPickerButton'}
-				onClick={onEmojiPickerButtonClick}
-				title={'Pick an emoji'}
-			>🙂
-			</button>
-
-			{(() => {
-				if (props.isEmojiPickerOpened) {
-					return <EmojiPicker
-						skinTonesDisabled={true}
-						defaultSkinTone={SkinTones.NEUTRAL}
-						emojiStyle={EmojiStyle.NATIVE}
-						previewConfig={{ showPreview: false }}
-						onEmojiClick={onEmojiClick}
-						width={'100%'}
-					/>;
-				} else {
-					return null;
-				}
-			})()}
 		</div>
 	);
 }
