+++ src/components/common/ChatInputEmojiPicker.css
@@ -1,0 +1,5 @@
+.pickerWrapper {
+    position: absolute;
+    width: 100%;
+    bottom: 0;
+}
