+++ src/state/reducer.ts
@@ -633,15 +633,6 @@ const reducer: Reducer<State> = (
 				}
 			};
 
-		case ActionTypes.EmojiPickerToggle:
-			return {
-				...state,
-				online: {
-					...state.online,
-					isEmojiPickerOpened: action.isOpened
-				}
-			};
-
 		default:
 			return {
 				...state,
