+++ src/state/room/RoomState.ts
@@ -58,7 +58,6 @@ export default interface RoomState {
 	chat: {
 		isVisible: boolean;
 		isActive: boolean;
-		isEmojiPickerOpened: boolean;
 		mode: ChatMode;
 		message: string;
 		messages: ChatMessage[];
@@ -171,7 +170,6 @@ export const initialState: RoomState = {
 	chat: {
 		isVisible: false,
 		isActive: false,
-		isEmojiPickerOpened: false,
 		mode: ChatMode.Chat,
 		message: '',
 		messages: [],
